got the testing working

+++ k2/syscall.h
@@ -12,7 +12,7 @@ void Exception(uint64_t esr_el1);
 void Kill(int p);
 int KernelCreate(int priority, void (*function)(), int parent);
 int Create(int priority, void (*function)());
-int CreateArgs(int priority, void (*function)(), int8_t argsno, int64_t *args);
+int CreateArgs(int priority, void (*function)(), uint64_t argsno, uint64_t *args);
 void Schedule();
 
 int MyTid();
